tweak(format): rename ADT "supports" to "support"
This is in anticipation of a change to provide default "supports"
values. This will change the effect of setting the keyword from being
"here's a list of all the supported types of the transformer" to "here's
the list of types I want you to support". With the change form
exhaustive to restrictive, changing the key to "support" seems more apt.