feat(qualifiedtype): resolve types more robustly
In some cases, a module may be loaded but not available under Main. To
be exhaustive, we can check Base.loaded_modules for any packages with
the expected name. While we are at it, we can use isdefined to check for
the existence of a module/type without having to resort to error
catching.