feat: overhaul type/method reflection inspection
To clean up the code previously in _read, I've reconsidered the approach
taken to type/method inspection via mostly-private reflection.

It would benefit from a little more work, but I'm much happier with this
new approach. It's less hacky even if it looks more complicated (and
still a bit hacky).