feat!: Merge the AbstractDataTransformers
It occurred to me that while I did have a reason originally for
separating these types out (IIRC something with the now seemingly unused
extra parameter with the DataStorage), that might have vanished in the
ongoing renovations and so I could try replacing the
AbstractDataTransformer type and its three concrete subtypes with a
single concrete DataTransformer type and an extra type parameter.

Nothing seems to have broken, so let's give it a shot 🙂, hopefully it
will allow for more efficient codegen.