tweak(Core): make DataCollection mutable
As nice as immutability is, I'm about to want to extend the structure by
turning .path into a tuple with the path as well as the mtime of the
source file. When writing new versions of the collection, it really
makes much more sense to be able to just update this new mtime field.

The swapping of a DataCollection in the STACK as a way to make
mutation-type operations seem to *sometimes* work is also a bit dodgy
upon reflection. These can (mostly) be cleaned up without making use of
the new mutability, but it makes sense to me to clean them up at the
same time.